moved all duplicated code from XPathContainsAssertion and XQueryContainsAssertion to XBaseContainsAssertion
